requirements.txt
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-monitor==6.0.2
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-storage==22.1.1
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
boto3==1.37.26
boto3-stubs==1.37.27
boto3-stubs-full==1.37.27
botocore==1.37.26
botocore-stubs==1.37.27
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
fonttools==4.57.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
microsoft-python-type-stubs @ git+https://github.com/microsoft/python-type-stubs.git@dcb5c8c2a61a2e4653b4becc9f20f6c42858a701
msal==1.32.0
msal-extensions==1.3.1
mypy-boto3-ec2==1.37.28
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
soupsieve==2.7
types-awscrt==0.25.7
types-s3transfer==0.11.4
typing_extensions==4.13.1
urllib3==2.3.0
